+++ src/dis6/EntityStatePdu.cpp
@@ -290,7 +290,7 @@ int EntityStatePdu::getMarshalledSize() const
    marshalSize = marshalSize + _marking.getMarshalledSize();  // _marking
    marshalSize = marshalSize + 4;  // _capabilities
 
-   for(unsigned long idx=0; idx < _articulationParameters.size(); idx++)
+   for(unsigned long long idx=0; idx < _articulationParameters.size(); idx++)
    {
         ArticulationParameter listElement = _articulationParameters[idx];
         marshalSize = marshalSize + listElement.getMarshalledSize();
