+++ src/dis6/UaPdu.cpp
@@ -277,21 +277,21 @@ int UaPdu::getMarshalledSize() const
    marshalSize = marshalSize + 1;  // _numberOfAPAs
    marshalSize = marshalSize + 1;  // _numberOfUAEmitterSystems
 
-   for(unsigned long idx=0; idx < _shaftRPMs.size(); idx++)
+   for(unsigned long long idx=0; idx < _shaftRPMs.size(); idx++)
    {
         ShaftRPMs listElement = _shaftRPMs[idx];
         marshalSize = marshalSize + listElement.getMarshalledSize();
     }
 
 
-   for(unsigned long idx=0; idx < _apaData.size(); idx++)
+   for(unsigned long long idx=0; idx < _apaData.size(); idx++)
    {
         ApaData listElement = _apaData[idx];
         marshalSize = marshalSize + listElement.getMarshalledSize();
     }
 
 
-   for(unsigned long idx=0; idx < _emitterSystems.size(); idx++)
+   for(unsigned long long idx=0; idx < _emitterSystems.size(); idx++)
    {
         AcousticEmitterSystemData listElement = _emitterSystems[idx];
         marshalSize = marshalSize + listElement.getMarshalledSize();
